etl_project/requirements.txt
pyspark==3.4.0
pandas==2.1.1
numpy==1.25.1
pyyaml==6.0
matplotlib==3.7.1
seaborn==0.12.2
prometheus_client==0.16.0
psutil==5.9.3
python-dotenv==1.0.0
apache-airflow==2.6.1
s3fs==2023.9.0
hadoop==3.3.1
openpyxl==3.0.10
pyarrow==12.0.0
sqlalchemy==2.0.17
psycopg2==2.9.5
pymysql==1.0.2